control_node/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
cryptography==42.0.5
PyYAML==6.0.1
requests==2.31.0
python-dateutil==2.9.0
colorama==0.4.6
tabulate==0.9.0
